feat(Home,Updateorg): Update org phone changed to the way backend asked and fixed some minor bugs

+++ src/@core/feature/DeleteOrg/ui/index.tsx
@@ -3,6 +3,7 @@ import { useLang } from '@/@core/shared/hooks/useLang'
 import { Img, Tooltip } from '@chakra-ui/react'
 import { FC } from 'react'
 import { toast } from 'react-toastify'
+import Swal from 'sweetalert2'
 
 type Props = {
   id: string
@@ -13,13 +14,25 @@ export const DeleteOrg: FC<Props> = props => {
   const { t } = useLang()
 
   const onDelete = async () => {
-    const res = await deleteOrganization(id)
+    Swal.fire({
+      title: t('delete'),
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: 'crimson',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: t('yes'),
+      cancelButtonText: t('no')
+    }).then(async res => {
+      if (res.isConfirmed) {
+        const res = await deleteOrganization(id)
 
-    if (!res) return null
+        if (!res) return null
 
-    if (res.status === 204) {
-      toast.success(t('deleted'), { position: 'bottom-right' })
-    }
+        if (res.status === 204) {
+          toast.success(t('deleted'), { position: 'bottom-right' })
+        }
+      }
+    })
   }
 
   return (
